requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
openai==1.93.0
anthropic==0.57.1
google-genai==0.8.0
celery==5.3.0
redis==5.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
jsonschema==4.20.0
beautifulsoup4==4.12.0
